requirements.txt
flask
flask-pymongo
flask-bcrypt
flask-cors
flask-session
dnspython